[ADD] ViewController에서 View 분리
